fix: restore output capture for in-process lensing
