+++ lib/Lens.js
@@ -1,10 +1,14 @@
 const axios = require('axios');
+const TOML = require('@iarna/toml');
+const handlebars = require('handlebars');
+const mkdirp = require('mz-modules/mkdirp');
+const shellParse = require('shell-quote-word');
+const squish = require('object-squish');
 
-
+const Repo = require('./Repo.js');
 const Configurable = require('./Configurable.js');
 const SpecObject = require('./SpecObject.js');
 const Studio = require('./Studio.js');
-
 
 const logger = require('./logger');
 
@@ -171,10 +175,112 @@ class Lens extends Configurable {
         };
     }
 
-    async execute (specHash) {
-        const studio = await Studio.get(this.workspace.getRepo().gitDir);
+    async executeSpec (specHash, options) {
+        return Lens.executeSpec(specHash, {...options, repo: this.workspace.getRepo()});
+    }
 
-        return studio.holoLensExec(specHash);
+    static async executeSpec (specHash, { refresh=false, save=true, repo=null }) {
+
+        // return studio.holoLensExec(specHash);
+
+        // TODO: run in studio if not in one?
+        // TODO: delegate some of this to Studio
+
+
+        // load holorepo
+        if (!repo) {
+            repo = await Repo.getFromEnvironment();
+        }
+
+        const git = await repo.getGit();
+
+
+        // ensure the rest runs inside a studio environment
+        const studio = await Studio.get(repo.gitDir);
+
+        if (!studio.isLocal()) {
+            return studio.holoLensExec(specHash);
+        }
+
+
+        // check for existing build
+        const specRef = SpecObject.buildRef('lens', specHash);
+        if (!refresh) {
+            const existingBuildHash = await repo.resolveRef(specRef);
+
+            if (existingBuildHash) {
+                return existingBuildHash;
+            }
+        }
+
+
+        // load spec
+        const specToml = await git.catFile({ p: true }, specHash);
+        const {
+            holospec: {
+                lens: spec
+            }
+        } = TOML.parse(specToml);
+
+
+        // assign scratch directory
+        const scratchPath = `${process.env.HOLO_SCRATCH||'/hab/cache/hololens'}/${spec.package.split('/').slice(0, 2).join('/')}`;
+        await mkdirp(scratchPath);
+
+
+        // install lens package
+        const hab = await Studio.getHab();
+        logger.info(`installing lens package: ${spec.package}`);
+        await hab.pkg('install', spec.package);
+
+
+        // prepare command and environment
+        const command = handlebars.compile(spec.command)(spec);
+        const env = Object.assign(
+            squish({
+                hololens: { ...spec, spec: specHash }
+            }, { seperator: '_', modifyKey: 'uppercase' }),
+            {
+                GIT_DIR: repo.gitDir,
+                GIT_WORK_TREE: scratchPath,
+                GIT_INDEX_FILE: `${scratchPath}.index`,
+                DEBUG: process.env.DEBUG || ''
+            }
+        );
+        logger.debug('lens environment:', env);
+
+
+        // spawn process and log STDERR
+        logger.info(`executing lens command: ${command}`);
+        const lensProcess = await hab.pkg('exec', spec.package, ...shellParse(command), {
+            $env: env,
+            $cwd: scratchPath,
+            $spawn: true
+        });
+
+        lensProcess.stderr.on('data', data => {
+            data.toString().trim().split(/\n/).forEach(
+                line => logger.info(`lens: ${line}`)
+            )
+        });
+
+
+        // process output
+        const lensedTreeHash = await lensProcess.captureOutputTrimmed();
+
+        if (!git.isHash(lensedTreeHash)) {
+            throw new Error(`lens "${command}" did not return hash: ${lensedTreeHash}`);
+        }
+
+
+        // save ref to accelerate next projection
+        if (save) {
+            await git.updateRef(specRef, lensedTreeHash);
+        }
+
+
+        // return tree hash
+        return lensedTreeHash;
     }
 
 }
