+++ lib/Projection.js
@@ -171,7 +171,7 @@ class Projection {
             if (outputTreeHash) {
                 logger.info(`found existing output tree matching holospec(${specHash})`);
             } else {
-                outputTreeHash = await lens.execute(specHash);
+                outputTreeHash = await lens.executeSpec(specHash);
             }
