+++ lib/Studio.js
@@ -253,10 +253,6 @@ class Studio {
     }
 
     async holoLensExec(spec) {
-        if (this.isLocal()) {
-            return require('../commands/lens/exec.js').handler({ spec });
-        }
-
         return this.holoExec('lens', 'exec', spec);
     }
